fix: All raw datasets fixed and tests passing

+++ src/datasets/datasets.controller.ts
@@ -374,7 +374,7 @@ export class DatasetsController {
       typeof where === "string" || (where as unknown) instanceof String
         ? JSON.parse(where)
         : where;
-    //console.log("Where : " + JSON.stringify(whereFilters));
+
     return this.datasetsService.count(whereFilters);
   }
 
@@ -533,7 +533,6 @@ export class DatasetsController {
   // POST /datasets/:id/origdatablocks
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => {
-    //console.log('dataset/<id>/origdatablock', ability.can(Action.Create, OrigDatablock));
     return ability.can(Action.Create, OrigDatablock);
   })
   @UseInterceptors(
@@ -619,10 +618,28 @@ export class DatasetsController {
     @Param("fk") origDatablockId: string,
     @Body() updateOrigdatablockDto: UpdateOrigDatablockDto,
   ): Promise<OrigDatablock | null> {
-    return this.origDatablocksService.update(
-      { _id: origDatablockId, datasetId },
-      updateOrigdatablockDto,
-    );
+    const dataset = await this.datasetsService.findOne({ pid: datasetId });
+    const origDatablockBeforeUpdate = await this.origDatablocksService.findOne({
+      _id: origDatablockId,
+    });
+    if (dataset && origDatablockBeforeUpdate) {
+      const origDatablock = await this.origDatablocksService.update(
+        { _id: origDatablockId, datasetId },
+        updateOrigdatablockDto,
+      );
+      if (origDatablock) {
+        await this.datasetsService.findByIdAndUpdate(datasetId, {
+          size:
+            dataset.size - origDatablockBeforeUpdate.size + origDatablock.size,
+          numberOfFiles:
+            dataset.numberOfFiles -
+            origDatablockBeforeUpdate.dataFileList.length +
+            origDatablock.dataFileList.length,
+        });
+        return origDatablock;
+      }
+    }
+    return null;
   }
 
   // DELETE /datasets/:id/origdatablocks/:fk
@@ -682,8 +699,11 @@ export class DatasetsController {
       };
       const datablock = await this.datablocksService.create(createDatablock);
       await this.datasetsService.findByIdAndUpdate(id, {
-        packedSize: datablock.packedSize,
-        numberOfFilesArchived: datablock.dataFileList.length,
+        packedSize: dataset.packedSize + datablock.packedSize,
+        numberOfFilesArchived:
+          dataset.numberOfFilesArchived + datablock.dataFileList.length,
+        size: dataset.size + datablock.size,
+        numberOfFiles: dataset.numberOfFiles + datablock.dataFileList.length,
       });
       return datablock;
     }
@@ -710,16 +730,33 @@ export class DatasetsController {
     @Param("fk") datablockId: string,
     @Body() updateDatablockDto: UpdateDatablockDto,
   ): Promise<Datablock | null> {
-    const datablock = await this.datablocksService.update(
-      { _id: datablockId, datasetId },
-      updateDatablockDto,
-    );
-    if (datablock) {
-      await this.datasetsService.findByIdAndUpdate(datasetId, {
-        packedSize: datablock.packedSize,
-        numberOfFilesArchived: datablock.dataFileList.length,
-      });
-      return datablock;
+    const dataset = await this.datasetsService.findOne({ pid: datasetId });
+    const datablockBeforeUpdate = await this.datablocksService.findOne({
+      _id: datablockId,
+    });
+    if (dataset && datablockBeforeUpdate) {
+      const datablock = await this.datablocksService.update(
+        { _id: datablockId, datasetId },
+        updateDatablockDto,
+      );
+      if (datablock) {
+        await this.datasetsService.findByIdAndUpdate(datasetId, {
+          packedSize:
+            dataset.packedSize -
+            datablockBeforeUpdate.packedSize +
+            datablock.packedSize,
+          numberOfFilesArchived:
+            dataset.numberOfFilesArchived -
+            datablockBeforeUpdate.dataFileList.length +
+            datablock.dataFileList.length,
+          size: dataset.size - datablockBeforeUpdate.size + datablock.size,
+          numberOfFiles:
+            dataset.numberOfFiles -
+            datablockBeforeUpdate.dataFileList.length +
+            datablock.dataFileList.length,
+        });
+        return datablock;
+      }
     }
     return null;
   }
@@ -732,9 +769,36 @@ export class DatasetsController {
     @Param("id") datasetId: string,
     @Param("fk") datablockId: string,
   ): Promise<unknown> {
-    return this.datablocksService.remove({
-      _id: datablockId,
-      datasetId,
-    });
+    const dataset = await this.datasetsService.findOne({ pid: datasetId });
+    if (dataset) {
+      // remove datablock
+      const res = await this.datablocksService.remove({
+        _id: datablockId,
+        datasetId,
+      });
+      // all the remaining datablocks for this dataset
+      const remainingDatablocks = await this.datablocksService.findAll({
+        datasetId: datasetId,
+      });
+      // update dataset size and files number
+      const updateDatasetDto: UpdateDatasetDto = {
+        packedSize: remainingDatablocks.reduce((a, b) => a + b.packedSize, 0),
+        numberOfFilesArchived: remainingDatablocks.reduce(
+          (a, b) => a + b.dataFileList.length,
+          0,
+        ),
+        size: remainingDatablocks.reduce((a, b) => a + b.size, 0),
+        numberOfFiles: remainingDatablocks.reduce(
+          (a, b) => a + b.dataFileList.length,
+          0,
+        ),
+      };
+      await this.datasetsService.findByIdAndUpdate(
+        dataset.pid,
+        updateDatasetDto,
+      );
+      return res;
+    }
+    return null;
   }
 }
